Split harness into modules
