+++ NFT/tests/functions/burn.rs
@@ -1,0 +1,105 @@
+use crate::utils::{
+    abi_calls::{balance_of, burn, constructor, mint, total_supply},
+    test_helpers::setup,
+    Identity,
+};
+use fuels::signers::Signer;
+
+mod success {
+
+    use super::*;
+
+    #[tokio::test]
+    async fn burns() {
+        let (deploy_wallet, owner1, _owner2) = setup().await;
+
+        // constructor(false, &deploy_wallet.contract, &Option::None(), 1).await;
+        let admin = Identity::Address(owner1.wallet.address());
+        constructor(true, &deploy_wallet.contract, &admin, 1).await;
+
+        let minter = Identity::Address(owner1.wallet.address());
+        mint(1, &owner1.contract, &minter).await;
+
+        assert_eq!(total_supply(&owner1.contract).await, 1);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 1);
+
+        burn(&owner1.contract, 0).await;
+
+        assert_eq!(total_supply(&owner1.contract).await, 0);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 0);
+    }
+
+    #[tokio::test]
+    async fn burns_multiple() {
+        let (deploy_wallet, owner1, _owner2) = setup().await;
+
+        // constructor(false, &deploy_wallet.contract, &Option::None(), 1).await;
+        let admin = Identity::Address(owner1.wallet.address());
+        constructor(true, &deploy_wallet.contract, &admin, 4).await;
+
+        let minter = Identity::Address(owner1.wallet.address());
+        mint(4, &owner1.contract, &minter).await;
+
+        assert_eq!(total_supply(&owner1.contract).await, 4);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 4);
+
+        burn(&owner1.contract, 0).await;
+        assert_eq!(total_supply(&owner1.contract).await, 3);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 3);
+
+        burn(&owner1.contract, 1).await;
+        assert_eq!(total_supply(&owner1.contract).await, 2);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 2);
+
+        burn(&owner1.contract, 2).await;
+        assert_eq!(total_supply(&owner1.contract).await, 1);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 1);
+
+        burn(&owner1.contract, 3).await;
+        assert_eq!(total_supply(&owner1.contract).await, 0);
+        assert_eq!(balance_of(&owner1.contract, &minter).await, 0);
+    }
+}
+
+mod reverts {
+
+    use super::*;
+
+    #[tokio::test]
+    #[should_panic(expected = "Revert(42)")]
+    async fn panics_when_token_owner_does_not_exist() {
+        let (_deploy_wallet, owner1, _owner2) = setup().await;
+
+        burn(&owner1.contract, 0).await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "Revert(42)")]
+    async fn panics_when_token_does_not_exist() {
+        let (deploy_wallet, owner1, _owner2) = setup().await;
+
+        // constructor(false, &deploy_wallet.contract, &Option::None(), 1).await;
+        let admin = Identity::Address(owner1.wallet.address());
+        constructor(true, &deploy_wallet.contract, &admin, 1).await;
+
+        let minter = Identity::Address(owner1.wallet.address());
+        mint(1, &owner1.contract, &minter).await;
+
+        burn(&owner1.contract, 1).await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "Revert(42)")]
+    async fn panics_when_sender_is_not_owner() {
+        let (deploy_wallet, owner1, owner2) = setup().await;
+
+        // constructor(false, &deploy_wallet.contract, &Option::None(), 1).await;
+        let admin = Identity::Address(owner1.wallet.address());
+        constructor(true, &deploy_wallet.contract, &admin, 1).await;
+
+        let minter = Identity::Address(owner1.wallet.address());
+        mint(1, &owner1.contract, &minter).await;
+
+        burn(&owner2.contract, 0).await;
+    }
+}
