+++ NFT/tests/functions/mod.rs
@@ -1,0 +1,15 @@
+mod admin;
+mod approve;
+mod approved;
+mod balance_of;
+mod burn;
+mod constructor;
+mod is_approved_for_all;
+mod max_supply;
+mod meta_data;
+mod mint;
+mod owner_of;
+mod set_admin;
+mod set_approval_for_all;
+mod total_supply;
+mod transfer_from;
